feat: DOCS-165 - Create python script to configure multi instance docs-editor
The script called carbonio-docs-routes-generator.py:
 - Find all the instances of a service
 - Auto-generate the service-resolver.json based on the instances already installed and the one being installed
 - Auto-generate the service-router.json based on the instances already installed and the one being installed
 - Update the sidecar configuration (service.hcl) in order to associate the meta tag with the service id generated

The carbonio-docs-editor-template.hcl has a new tag meta that contains two variables substituted by the script during the setup. The first variable represents the keyword used to identify the id of the service in a http request, the second one the UUID of the service also generated during the setup.

Approved-by: Daniele Giachetto
Approved-by: Gianluca Boiano
Approved-by: Domenico Stragliotto